Rename domain objects
