+++ src/main/java/com/dnastack/discovery/registry/repository/ServiceNodeRepository.java
@@ -1,11 +1,11 @@
 package com.dnastack.discovery.registry.repository;
 
-import com.dnastack.discovery.registry.domain.ServiceNodeEntity;
+import com.dnastack.discovery.registry.domain.ServiceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServiceNodeRepository extends JpaRepository<ServiceNodeEntity, String>, QuerydslPredicateExecutor<ServiceNodeEntity> {
+public interface ServiceNodeRepository extends JpaRepository<ServiceEntity, String>, QuerydslPredicateExecutor<ServiceEntity> {
 
 }
